Move test_push from mod.rs to push.rs
Towards #290
