+++ src/sync/mod.rs
@@ -760,602 +760,6 @@ mod tests {
         }
     }
 
-    #[async_std::test]
-    async fn test_push() {
-        let store = dag::Store::new(Box::new(MemStore::new()));
-        let mut chain: Chain = vec![];
-        add_genesis(&mut chain, &store).await;
-        add_snapshot(&mut chain, &store, Some(vec![str!("foo"), str!("bar")])).await;
-        // chain[2] is an index change
-        add_index_change(&mut chain, &store).await;
-        let starting_num_commits = chain.len();
-        let base_snapshot = &chain[1];
-        let (base_last_mutation_id, base_server_state_id) =
-            Commit::snapshot_meta_parts(base_snapshot).unwrap();
-
-        let sync_id = str!("sync_id");
-        let client_id = str!("test_client_id");
-        let data_layer_auth = str!("data_layer_auth");
-
-        // Push
-        let batch_push_url = str!("batch_push_url");
-        let client_view_url = str!("client_view_url");
-
-        // Pull
-        let diff_server_url = str!("diff_server_url");
-        let diff_server_auth = str!("diff_server_auth");
-
-        let good_client_view_info = ClientViewInfo {
-            http_status_code: 200,
-            error_message: str!(""),
-        };
-        let good_pull_resp = PullResponse {
-            state_id: str!("new_state_id"),
-            last_mutation_id: 10,
-            patch: vec![
-                Operation {
-                    op: str!("replace"),
-                    path: str!(""),
-                    value_string: str!("{}"),
-                },
-                Operation {
-                    op: str!("add"),
-                    path: str!("/new"),
-                    value_string: str!("\"value\""),
-                },
-            ],
-            client_view_info: good_client_view_info.clone(),
-        };
-        let good_pull_resp_value_map = map!("/new" => "\"value\"");
-
-        struct ExpCommit<'a> {
-            state_id: String,
-            last_mutation_id: u64,
-            value_map: HashMap<&'a str, &'a str>,
-            indexes: Vec<String>,
-        }
-
-        struct Case<'a> {
-            pub name: &'a str,
-
-            // Push expectations.
-            pub num_pending_mutations: u32,
-            pub exp_push_req: Option<push::BatchPushRequest>,
-            pub push_result: Option<Result<push::BatchPushResponse, String>>,
-
-            // Pull expectations.
-            pub exp_pull_req: PullRequest,
-            pub pull_result: Result<PullResponse, String>,
-
-            // BeginPull expectations.
-            pub exp_err: Option<&'a str>,
-            pub exp_new_sync_head: Option<ExpCommit<'a>>,
-        }
-        let cases: Vec<Case> = vec![
-            Case {
-                name: "0 pending, pulls new state -> beginpull succeeds w/synchead set",
-                num_pending_mutations: 0,
-                exp_push_req: None,
-                push_result: None,
-                exp_pull_req: PullRequest {
-                    client_view_auth: data_layer_auth.clone(),
-                    client_view_url: client_view_url.clone(),
-                    client_id: client_id.clone(),
-                    base_state_id: base_server_state_id.clone(),
-                    last_mutation_id: base_last_mutation_id,
-                    version: 3,
-                },
-                pull_result: Ok(good_pull_resp.clone()),
-                exp_err: None,
-                exp_new_sync_head: Some(ExpCommit {
-                    state_id: str!("new_state_id"),
-                    last_mutation_id: 10,
-                    value_map: good_pull_resp_value_map.clone(),
-                    indexes: vec![2.to_string()],
-                }),
-            },
-            Case {
-                name: "1 pending, 0 mutations to replay, pulls new state -> beginpull succeeds w/synchead set",
-                num_pending_mutations: 1,
-                exp_push_req: Some(push::BatchPushRequest {
-                    client_id: client_id.clone(),
-                    mutations: vec![
-                        push::Mutation {
-                            id: 2,
-                            name: "mutator_name_3".to_string(),
-                            args: json!([3]),
-                        },
-                    ],
-                }),
-                push_result: Some(Ok(push::BatchPushResponse {})),
-                exp_pull_req: PullRequest {
-                    client_view_auth: data_layer_auth.clone(),
-                    client_view_url: client_view_url.clone(),
-                    client_id: client_id.clone(),
-                    base_state_id: base_server_state_id.clone(),
-                    last_mutation_id: base_last_mutation_id,
-                    version: 3,
-                },
-                pull_result: Ok(PullResponse {
-                    last_mutation_id: 2,
-                    ..good_pull_resp.clone()
-                }),
-                exp_err: None,
-                exp_new_sync_head: Some(ExpCommit {
-                    state_id: str!("new_state_id"),
-                    last_mutation_id: 2,
-                    value_map: good_pull_resp_value_map.clone(),
-                    indexes: vec![2.to_string(), 4.to_string()],
-                }),
-            },
-            Case {
-                name: "1 pending, 1 mutations to replay, pulls new state -> beginpull succeeds w/synchead set",
-                num_pending_mutations: 1,
-                exp_push_req: Some(push::BatchPushRequest {
-                    client_id: client_id.clone(),
-                    mutations: vec![
-                        push::Mutation {
-                            id: 2,
-                            name: "mutator_name_3".to_string(),
-                            args: json!([3]),
-                        },
-                    ],
-                }),
-                push_result: Some(Ok(push::BatchPushResponse {})),
-                exp_pull_req: PullRequest {
-                    client_view_auth: data_layer_auth.clone(),
-                    client_view_url: client_view_url.clone(),
-                    client_id: client_id.clone(),
-                    base_state_id: base_server_state_id.clone(),
-                    last_mutation_id: base_last_mutation_id,
-                    version: 3,
-                },
-                pull_result: Ok(PullResponse {
-                    last_mutation_id: 1,
-                    ..good_pull_resp.clone()
-                }),
-                exp_err: None,
-                exp_new_sync_head: Some(ExpCommit {
-                    state_id: str!("new_state_id"),
-                    last_mutation_id: 1,
-                    value_map: good_pull_resp_value_map.clone(),
-                    indexes: vec![2.to_string()],
-                }),
-            },
-            Case {
-                name: "2 pending, 0 to replay, push succeeds, pulls new state -> beginpull succeeds w/synchead set",
-                num_pending_mutations: 2,
-                exp_push_req: Some(push::BatchPushRequest {
-                    client_id: client_id.clone(),
-                    mutations: vec![
-                        // These mutations aren't actually added to the chain until the test
-                        // case runs, but we happen to know how they are created by the db
-                        // test helpers so we use that knowledge here.
-                        push::Mutation {
-                            id: 2,
-                            name: "mutator_name_3".to_string(),
-                            args: json!([3]),
-                        },
-                        push::Mutation {
-                            id: 3,
-                            name: "mutator_name_5".to_string(),
-                            args: json!([5]),
-                        },
-                    ],
-                }),
-                push_result: Some(Ok(push::BatchPushResponse {})),
-                exp_pull_req: PullRequest {
-                    client_view_auth: data_layer_auth.clone(),
-                    client_view_url: client_view_url.clone(),
-                    client_id: client_id.clone(),
-                    base_state_id: base_server_state_id.clone(),
-                    last_mutation_id: base_last_mutation_id,
-                    version: 3,
-                },
-                pull_result: Ok(good_pull_resp.clone()),
-                exp_err: None,
-                exp_new_sync_head: Some(ExpCommit {
-                    state_id: str!("new_state_id"),
-                    last_mutation_id: 10,
-                    value_map: good_pull_resp_value_map.clone(),
-                    indexes: vec![2.to_string(), 4.to_string(), 6.to_string()],
-                }),
-            },
-            Case {
-                name: "2 pending, 1 to replay, push succeeds, pulls new state -> beginpull succeeds w/synchead set",
-                num_pending_mutations: 2,
-                exp_push_req: Some(push::BatchPushRequest {
-                    client_id: client_id.clone(),
-                    mutations: vec![
-                        // These mutations aren't actually added to the chain until the test
-                        // case runs, but we happen to know how they are created by the db
-                        // test helpers so we use that knowledge here.
-                        push::Mutation {
-                            id: 2,
-                            name: "mutator_name_3".to_string(),
-                            args: json!([3]),
-                        },
-                        push::Mutation {
-                            id: 3,
-                            name: "mutator_name_5".to_string(),
-                            args: json!([5]),
-                        },
-                    ],
-                }),
-                push_result: Some(Ok(push::BatchPushResponse {})),
-                exp_pull_req: PullRequest {
-                    client_view_auth: data_layer_auth.clone(),
-                    client_view_url: client_view_url.clone(),
-                    client_id: client_id.clone(),
-                    base_state_id: base_server_state_id.clone(),
-                    last_mutation_id: base_last_mutation_id,
-                    version: 3,
-                },
-                pull_result: Ok(PullResponse {
-                    last_mutation_id: 2,
-                    ..good_pull_resp.clone()
-                }),
-                exp_err: None,
-                exp_new_sync_head: Some(ExpCommit {
-                    state_id: str!("new_state_id"),
-                    last_mutation_id: 2,
-                    value_map: good_pull_resp_value_map.clone(),
-                    indexes: vec![2.to_string(), 4.to_string()],
-                }),
-            },
-            Case {
-                name: "2 mutations to push, push errors, pulls new state -> beginpull succeeds w/synchead set",
-                num_pending_mutations: 2,
-                exp_push_req: Some(push::BatchPushRequest {
-                    client_id: client_id.clone(),
-                    mutations: vec![
-                        // These mutations aren't actually added to the chain until the test
-                        // case runs, but we happen to know how they are created by the db
-                        // test helpers so we use that knowledge here.
-                        push::Mutation {
-                            id: 2,
-                            name: "mutator_name_3".to_string(),
-                            args: json!([3]),
-                        },
-                        push::Mutation {
-                            id: 3,
-                            name: "mutator_name_5".to_string(),
-                            args: json!([5]),
-                        },
-                    ],
-                }),
-                push_result: Some(Err(str!("FetchNotOk(500)"))),
-                exp_pull_req: PullRequest {
-                    client_view_auth: data_layer_auth.clone(),
-                    client_view_url: client_view_url.clone(),
-                    client_id: client_id.clone(),
-                    base_state_id: base_server_state_id.clone(),
-                    last_mutation_id: base_last_mutation_id,
-                    version: 3,
-                },
-                pull_result: Ok(good_pull_resp.clone()),
-                exp_err: None,
-                exp_new_sync_head: Some(ExpCommit {
-                    state_id: str!("new_state_id"),
-                    last_mutation_id: 10,
-                    value_map: good_pull_resp_value_map.clone(),
-                    indexes: vec![2.to_string(), 4.to_string(), 6.to_string()],
-                }),
-            },
-
-            // TODO: add test for same state id but later mutation id. Right now we treat
-            // it as a nop because the state id does not change, but probably we should treat
-            // it like success and complete the sync. Current behavior mirrors the (probably
-            // incorrect?) go behavior.
-            Case {
-                name: "nop push, pulls same state -> beginpull succeeds with no synchead",
-                num_pending_mutations: 0,
-                exp_push_req: None,
-                push_result: None,
-                exp_pull_req: PullRequest {
-                    client_view_auth: data_layer_auth.clone(),
-                    client_view_url: client_view_url.clone(),
-                    client_id: client_id.clone(),
-                    base_state_id: base_server_state_id.clone(),
-                    last_mutation_id: base_last_mutation_id,
-                    version: 3,
-                },
-                pull_result: Ok(PullResponse {
-                    state_id: base_server_state_id.clone(),
-                    last_mutation_id: base_last_mutation_id,
-                    patch: vec![],
-                    client_view_info: good_client_view_info.clone(),
-                }),
-                exp_err: None,
-                exp_new_sync_head: None,
-            },
-            Case {
-                name: "nop push, pulls new state w/lesser mutation id -> beginpull errors",
-                num_pending_mutations: 0,
-                exp_push_req: None,
-                push_result: None,
-                exp_pull_req: PullRequest {
-                    client_view_auth: data_layer_auth.clone(),
-                    client_view_url: client_view_url.clone(),
-                    client_id: client_id.clone(),
-                    base_state_id: base_server_state_id.clone(),
-                    last_mutation_id: base_last_mutation_id,
-                    version: 3,
-                },
-                pull_result: Ok(PullResponse {
-                    last_mutation_id: 0,
-                    ..good_pull_resp.clone()
-                }),
-                exp_err: Some("TimeTravel"),
-                exp_new_sync_head: None,
-            },
-            Case {
-                name: "nop push, pulls new state w/empty state id -> beginpull errors",
-                num_pending_mutations: 0,
-                exp_push_req: None,
-                push_result: None,
-                exp_pull_req: PullRequest {
-                    client_view_auth: data_layer_auth.clone(),
-                    client_view_url: client_view_url.clone(),
-                    client_id: client_id.clone(),
-                    base_state_id: base_server_state_id.clone(),
-                    last_mutation_id: base_last_mutation_id,
-                    version: 3,
-                },
-                pull_result: Ok(PullResponse {
-                    state_id: str!(""),
-                    ..good_pull_resp.clone()
-                }),
-                exp_err: Some("MissingStateID"),
-                exp_new_sync_head: None,
-            },
-            Case {
-                name: "nop push, pull 500s -> beginpull errors",
-                num_pending_mutations: 0,
-                exp_push_req: None,
-                push_result: None,
-                exp_pull_req: PullRequest {
-                    client_view_auth: data_layer_auth.clone(),
-                    client_view_url: client_view_url.clone(),
-                    client_id: client_id.clone(),
-                    base_state_id: base_server_state_id.clone(),
-                    last_mutation_id: base_last_mutation_id,
-                    version: 3,
-                },
-                pull_result: Err(str!("FetchNotOk(500)")),
-                exp_err: Some("FetchNotOk(500)"),
-                exp_new_sync_head: None,
-            },
-        ];
-        for c in cases.iter() {
-            // Reset state of the store.
-            chain.truncate(starting_num_commits);
-            let w = store.write(LogContext::new()).await.unwrap();
-            w.set_head(
-                DEFAULT_HEAD_NAME,
-                Some(chain[chain.len() - 1].chunk().hash()),
-            )
-            .await
-            .unwrap();
-            w.set_head(SYNC_HEAD_NAME, None).await.unwrap();
-            w.commit().await.unwrap();
-            for _ in 0..c.num_pending_mutations {
-                add_local(&mut chain, &store).await;
-                add_index_change(&mut chain, &store).await;
-            }
-
-            // There was an index added after the snapshot, and one for each local commit.
-            // Here we scan to ensure that we get values when scanning using one of the
-            // indexes created. We do this because after calling begin_sync we check that
-            // the index no longer returns values, demonstrating that it was rebuilt.
-            if c.num_pending_mutations > 0 {
-                let dag_read = store.read(LogContext::new()).await.unwrap();
-                let read = db::OwnedRead::from_whence(
-                    db::Whence::Head(DEFAULT_HEAD_NAME.to_string()),
-                    dag_read,
-                )
-                .await
-                .unwrap();
-                use std::cell::RefCell;
-                let got = RefCell::new(false);
-
-                read.as_read()
-                    .scan(
-                        db::ScanOptions {
-                            prefix: Some(str!("")),
-                            start_secondary_key: None,
-                            start_key: None,
-                            start_exclusive: None,
-                            limit: None,
-                            index_name: Some(str!("2")),
-                        },
-                        |_: db::ScanResult<'_>| {
-                            *got.borrow_mut() = true;
-                        },
-                    )
-                    .await
-                    .unwrap();
-                assert!(*got.borrow(), "{}: expected values, got none", c.name);
-            }
-
-            // See explanation in FakePusher for why we do this dance with the push_result.
-            let (exp_push, push_resp, push_err) = match &c.push_result {
-                Some(Ok(resp)) => (true, Some(resp.clone()), None),
-                Some(Err(e)) => (true, None, Some(e.clone())),
-                None => (false, None, None),
-            };
-            let fake_pusher = FakePusher {
-                exp_push,
-                exp_push_req: c.exp_push_req.as_ref(),
-                exp_batch_push_url: &batch_push_url,
-                exp_batch_push_auth: &data_layer_auth,
-                exp_sync_id: &sync_id,
-                resp: push_resp,
-                err: push_err,
-            };
-
-            // See explanation in FakePuller for why we do this dance with the pull_result.
-            let (pull_resp, pull_err) = match &c.pull_result {
-                Ok(resp) => (Some(resp.clone()), None),
-                Err(e) => (None, Some(e.clone())),
-            };
-            let fake_puller = FakePuller {
-                exp_pull_req: &c.exp_pull_req,
-                exp_diff_server_url: &diff_server_url,
-                exp_diff_server_auth: &diff_server_auth,
-                exp_sync_id: &sync_id,
-                resp: pull_resp,
-                err: pull_err,
-            };
-
-            let begin_sync_req = BeginSyncRequest {
-                batch_push_url: batch_push_url.clone(),
-                client_view_url: client_view_url.clone(),
-                data_layer_auth: data_layer_auth.clone(),
-                diff_server_url: diff_server_url.clone(),
-                diff_server_auth: diff_server_auth.clone(),
-            };
-            let result = begin_sync(
-                &store,
-                LogContext::new(),
-                &fake_pusher,
-                &fake_puller,
-                begin_sync_req,
-                str!("test_client_id"),
-                sync_id.clone(),
-            )
-            .await;
-            let mut got_resp: Option<BeginSyncResponse> = None;
-            match c.exp_err {
-                None => {
-                    assert!(result.is_ok(), format!("{}: {:?}", c.name, result));
-                    got_resp = Some(result.unwrap());
-                }
-                Some(e) => assert!(to_debug(result.unwrap_err()).contains(e)),
-            };
-            let owned_read = store.read(LogContext::new()).await.unwrap();
-            let read = owned_read.read();
-            match &c.exp_new_sync_head {
-                None => {
-                    let got_head = read.get_head(SYNC_HEAD_NAME).await.unwrap();
-                    assert!(
-                        got_head.is_none(),
-                        format!(
-                            "{}: expected head to be None, was {}",
-                            c.name,
-                            got_head.unwrap()
-                        )
-                    );
-                    // In a nop sync we except Beginpull to succeed but sync_head will
-                    // be empty.
-                    if c.exp_err.is_none() {
-                        assert!(&got_resp.as_ref().unwrap().sync_head.is_empty());
-                    }
-                }
-                Some(exp_sync_head) => {
-                    let sync_head_hash = read.get_head(SYNC_HEAD_NAME).await.unwrap().unwrap();
-                    let sync_head =
-                        Commit::from_chunk(read.get_chunk(&sync_head_hash).await.unwrap().unwrap())
-                            .unwrap();
-                    let (got_last_mutation_id, got_server_state_id) =
-                        Commit::snapshot_meta_parts(&sync_head).unwrap();
-                    assert_eq!(exp_sync_head.last_mutation_id, got_last_mutation_id);
-                    assert_eq!(exp_sync_head.state_id, got_server_state_id);
-                    // Check the value is what's expected.
-                    let (_, _, map) = db::read_commit(
-                        db::Whence::Hash(sync_head.chunk().hash().to_string()),
-                        &read,
-                    )
-                    .await
-                    .unwrap();
-                    let mut got_value_map = map
-                        .iter()
-                        .map(|entry| (entry.key.to_vec(), entry.val.to_vec()))
-                        .collect::<Vec<(Vec<u8>, Vec<u8>)>>();
-                    got_value_map.sort_by(|a, b| a.0.cmp(&b.0));
-                    let exp_value_map = exp_sync_head
-                        .value_map
-                        .iter()
-                        .sorted()
-                        .map(|item| (item.0.as_bytes().to_vec(), item.1.as_bytes().to_vec()))
-                        .collect::<Vec<(Vec<u8>, Vec<u8>)>>();
-                    assert_eq!(exp_value_map.len(), got_value_map.len());
-
-                    // Check we have the expected index definitions.
-                    let indexes: Vec<String> = sync_head
-                        .indexes()
-                        .iter()
-                        .map(|i| i.definition.name.clone())
-                        .collect();
-                    assert_eq!(
-                        exp_sync_head.indexes.len(),
-                        indexes.len(),
-                        "{}: expected indexes {:?}, got {:?}",
-                        c.name,
-                        exp_sync_head.indexes,
-                        indexes
-                    );
-                    exp_sync_head
-                        .indexes
-                        .iter()
-                        .for_each(|i| assert!(indexes.contains(i)));
-
-                    // Check that we *don't* have old indexed values. The indexes should
-                    // have been rebuilt with a client view returned by the server that
-                    // does not include local= values. The check for len > 1 is because
-                    // the snapshot's index is not what we want; we want the first index
-                    // change's index ("2").
-                    if exp_sync_head.indexes.len() > 1 {
-                        let dag_read = store.read(LogContext::new()).await.unwrap();
-                        let read = db::OwnedRead::from_whence(
-                            db::Whence::Head(SYNC_HEAD_NAME.to_string()),
-                            dag_read,
-                        )
-                        .await
-                        .unwrap();
-                        read.as_read()
-                            .scan(
-                                db::ScanOptions {
-                                    prefix: Some(str!("")),
-                                    start_secondary_key: None,
-                                    start_key: None,
-                                    start_exclusive: None,
-                                    limit: None,
-                                    index_name: Some(str!("2")),
-                                },
-                                |sr: db::ScanResult<'_>| {
-                                    assert!(false, "{}: expected no values, got {:?}", c.name, sr);
-                                },
-                            )
-                            .await
-                            .unwrap();
-                    }
-
-                    assert_eq!(&sync_head_hash, &got_resp.as_ref().unwrap().sync_head);
-                }
-            }
-
-            // Check that SyncInfo is filled like we would expect.
-            if c.exp_err.is_none() {
-                let resp = got_resp.unwrap();
-                let got_push_info = resp.sync_info.batch_push_info;
-                match &c.push_result {
-                    Some(Ok(_)) => assert_eq!(200, got_push_info.unwrap().http_status_code),
-                    Some(Err(_)) => assert_eq!(500, got_push_info.unwrap().http_status_code),
-                    _ => (),
-                };
-
-                let got_client_view_info = resp.sync_info.client_view_info.as_ref().unwrap();
-                if !&c.pull_result.is_err() {
-                    assert_eq!(
-                        &c.pull_result.as_ref().unwrap().client_view_info,
-                        got_client_view_info
-                    );
-                }
-            }
-        }
-    }
-
     pub struct FakePusher<'a> {
         exp_push: bool,
         exp_push_req: Option<&'a push::BatchPushRequest>,
